build: upgrade report-changes.sh

+++ report-changes.sh
@@ -13,7 +13,7 @@ cd "$source_directory" || exit
 git add --all
 
 # Get the list of modified files in the git repository
-modified_files=$(git diff --name-only --cached)
+modified_files=$(git diff --name-only --staged)
 
 # Iterate over each modified file
 for file in $modified_files; do
@@ -37,18 +37,41 @@ for file in $modified_files; do
     # Exactly one match found
     target_file=${matching_files[0]}
   else
-    # No matches found, skip to the next file
-    echo "❌ No file $filename was found. Maybe it's because you renamed it ?"
-    continue
+    # No matches found
+    echo "❌ No file $filename was found."
+    renamed_files=$(git diff --name-status --staged | awk '/^R/ {print $3}')
+    created_files=$(git diff --name-status --staged | awk '/^A/ {print $2}')
+    
+    if [[ " ${renamed_files[@]} " =~ " ${filename} " ]]; then
+      echo "It has been renamed."
+      echo "Type the full path of the destination folder where the file to rename is located:"
+      read -r path
+      new_path="$path/$filename"
+      echo "Renaming file to: $new_path"
+      mv "$file" "$new_path"
+      target_file="$new_path"
+    elif [[ " ${created_files[@]} " =~ " ${filename} " ]]; then
+      echo "It has been created."
+      echo "Please provide the path where you want to create the file:"
+      read -r new_path
+      mkdir -p "$(dirname "$new_path")"  # Ensure the directory exists
+      touch "$new_path"
+      target_file="$new_path"
+    else
+      continue
+    fi
   fi
 
-  # Copy the content of the modified file to the target file
-  cp "$file" "$target_file"
-
-
-  echo "$target_file"
-  # Replace 'projectlocal' with '{{projectName}}' in the target file
-  sed -i '' 's/projectlocal/{{projectName}}/g' "$target_file"
-
-  echo "✅ Successfully pasted file $filename"
+  # Ensure the target file exists before attempting to copy
+  if [ -f "$file" ]; then
+    # Copy the content of the modified file to the target file
+    cp "$file" "$target_file"
+  
+    # Replace 'projectlocal' with '{{projectName}}' in the target file
+    sed -i '' 's/projectlocal/{{projectName}}/g' "$target_file"
+  
+    echo "✅ Successfully pasted file $filename"
+  else
+    echo "❌ Source file $file does not exist. Skipping."
+  fi
 done
